Allow type conversion from AttributeBackingType to int
To get around the compiler error
error: format specifies type int but the argument has type AttributeBackingType

Test: m .

Bug: b/312565119

+++ system/rust/src/gatt/ffi/gatt_shim.cc
@@ -76,7 +76,7 @@ void GattServerCallbacks::OnServerRead(uint16_t conn_id, uint32_t trans_id,
                          trans_id, addr.value(), attr_handle, offset, is_long));
       break;
     default:
-      LOG_ALWAYS_FATAL("Unexpected backing type %d", attr_type);
+      LOG_ALWAYS_FATAL("Unexpected backing type %d", (int)attr_type);
   }
 }
 
@@ -113,7 +113,7 @@ void GattServerCallbacks::OnServerWrite(
                          value.size()));
       break;
     default:
-      LOG_ALWAYS_FATAL("Unexpected backing type %hhu", attr_type);
+      LOG_ALWAYS_FATAL("Unexpected backing type %d", (int)attr_type);
   }
 }
